modify lints
